requirements.txt
base58==2.1.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cytoolz==0.12.2
ecdsa==0.18.0
eth-hash==0.5.2
eth-keys==0.4.0
eth-typing==3.5.2
eth-utils==2.3.1
futures==2.2.0
greenlet==3.0.3
idna==3.6
loguru==0.7.2
more-itertools==10.1.0
mysql-connector-python==8.2.0
protobuf==4.21.12
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pycparser==2.21
pycryptodome==3.19.0
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.31.0
scalecodec==1.2.7
six==1.16.0
SQLAlchemy==2.0.23
substrate-interface==1.7.4
toolz==0.12.0
typing_extensions==4.9.0
urllib3==2.1.0
websocket-client==1.7.0
xxhash==3.4.1
